requirements.txt
flask==3.1.1
numpy==1.26.4
pinecone==7.3.0
python-dotenv==1.1.1
sentence-transformers==5.0.0